requirements.txt
arrow==0.15.8
attrs==20.1.0
base==1.0.4
bcrypt==3.2.0
binaryornot==0.4.4
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cookiecutter==1.7.2
cryptography==3.0
devtools==0.6
dnspython==2.0.0
ecdsa==0.15
email-validator==1.1.1
Faker==4.1.2
fastapi==0.61.0
h11==0.10.0
httptools==0.1.1
ics==0.7
idna==2.10
iniconfig==1.0.1
Jinja2==2.11.2
jinja2-time==0.2.0
loguru==0.5.1
MarkupSafe==1.1.1
mongoengine==0.20.0
mongomock==3.20.0
more-itertools==8.4.0
names==0.3.0
packaging==20.4
passlib==1.7.2
pluggy==0.13.1
poyo==0.5.0
py==1.9.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.6.1
pymongo==3.11.0
pyparsing==2.4.7
pytest==6.0.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-jose==3.2.0
python-multipart==0.0.5
python-slugify==4.0.1
requests==2.24.0
rsa==4.6
sentinels==1.0.0
six==1.15.0
starlette==0.13.8
TatSu==5.5.0
text-unidecode==1.3
toml==0.10.1
urllib3==1.25.10
uvicorn==0.11.8
uvloop==0.14.0
wcwidth==0.2.5
websockets==8.1
